refactor: Nickname.jsx 추가로 메인코드 간소화
